Общий код вынесен в базовый класс
